+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Update docstrings for `call` and `first-cycle`.
 - Add `select-targets` function for selecting component names
   from a set of groups.
+- Allow weak refs that allow a component to wait for another
+  component without creating a hard dependency.
 
 ## v0.2.0 (2025-06-13)